[refactor]#108 coolsmsservice interface 분리
